HAYO AI AGENT ANDROID/requirements.txt
# HAYO AI Agent — Android Edition
# Core
chainlit>=1.1.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
python-dotenv>=1.0.0

# Memory (SQLite)
langgraph-checkpoint-sqlite>=2.0.0
aiosqlite>=0.20.0

# Web tools
duckduckgo-search>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# YouTube downloads (optional)
yt-dlp>=2024.1.0
